Changed score calculation

+++ CompareUtils.cpp
@@ -2,7 +2,7 @@
 
 #include <cmath>
 
-int CompareUtils::compare(const double& a, const double& b, double eps) {
+int CompareUtils::compare(const long double& a, const long double& b, long double eps) {
   if (std::abs(a - b) <= eps) return 0;
   return a < b ? -1 : +1;
 }
